+++ examples/Default.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
-import Flickity from '../src/index.js'
-import { images } from './images.js'
+import Flickity from '../src/index'
+import { images } from './images'
 
 export default function Default() {
   const [imgs, setImages] = useState(images)
