+++ examples/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="https://unpkg.com/flickity@2/dist/flickity.min.css">
     <title>React-Flickity-Component</title>
-    <script type="module" src="main.jsx"></script>
+    <script type="module" src="main.tsx"></script>
   </head>
   <body>
     <div id="root"></div>
