+++ vite.config.js
@@ -3,6 +3,7 @@ import { fileURLToPath } from 'url'
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import { viteCommonjs } from '@originjs/vite-plugin-commonjs'
+import dts from 'vite-plugin-dts'
 
 const __dirname = fileURLToPath(new URL('.', import.meta.url))
 
@@ -31,22 +32,28 @@ export default defineConfig(({ command, mode }) => {
     } else {
       return {
         ...base,
+        plugins: [
+          ...base.plugins,
+          dts({
+            include: ['src/**/*'],
+            exclude: ['**/*.test.*', '**/*.spec.*'],
+          }),
+        ],
         build: {
           sourcemap: true,
           lib: {
-            entry: resolve(__dirname, 'src/index.js'),
+            entry: resolve(__dirname, 'src/index.tsx'),
             name: 'ReactFlickityComponent',
             formats: ['es', 'umd'],
             fileName: (format) => `react-flickity-component.${format}.js`,
           },
           rollupOptions: {
-            external: ['react', 'react-dom', 'imagesloaded', 'prop-types', 'flickity'],        
+            external: ['react', 'react-dom', 'imagesloaded', 'flickity'],        
             output: {
               globals: {
                 react: 'React',
                 'react-dom': 'ReactDOM',
                 imagesloaded: 'imagesloaded',
-                'prop-types': 'PropTypes',
               },
             },
           },
